Use Ar container comparison when ElfContainer is not applicable.
Files with .a extension and "ar archive" in file type description
are treated as SharedLibrary and compared using readelf/objdump.
However, if the archive does not contain elf object files,
that approach will result in error from readelf (re. #849407).

It would be better to check if the .a file actually
contain elf files and if not, fall back to regular Ar-files
comparison. That should also help in case readelf/objdump is not
available.