Move 'container with one member' logic into separate class.
Create OneMenberContainer class, that incapsulates the logic of
container having only one member. It has the ability (get_nested_container)
to check if this only member is also a container, and in such case,
return this inner container.
Change Bzip2Container, GzipContainer and XzContainer to use this
class, avoiding DRY violations.